clustering/requirements.txt
face_recognition
scikit-learn
opencv-python-headless
# deepface
google-cloud-storage
numpy
google-auth
python-dotenv
# psutil
# facenet_pytorch
# hdbscan
# umap-learn
charset-normalizer
MarkupSafe
pillow